+++ platform/core/src/classes/MetadataProvider.js
@@ -441,7 +441,7 @@ class MetadataProvider {
   }
 
   getUIDsFromImageID(imageId) {
-    if (!imageId) throw new Error('Empty imageId');
+    if (!imageId) throw new Error('MetadataProvider::Empty imageId');
     // TODO: adding csiv here is not really correct. Probably need to use
     // metadataProvider.addImageIdToUIDs(imageId, {
     //   StudyInstanceUID,
